test: merge cli test files
