+++ lib/__tests__/cli.test.js
@@ -3,6 +3,7 @@
 'use strict';
 
 const path = require('path');
+const { promises: fs, existsSync } = require('fs');
 const stripAnsi = require('strip-ansi');
 
 const cli = require('../cli');
@@ -11,6 +12,12 @@ const replaceBackslashes = require('../testUtils/replaceBackslashes');
 
 const fixturesPath = (...elems) => replaceBackslashes(path.join(__dirname, 'fixtures', ...elems));
 const { buildCLI } = cli;
+
+const removeFile = async (filePath) => {
+	if (existsSync(filePath)) {
+		await fs.unlink(filePath);
+	}
+};
 
 jest.mock('../utils/getStdin', () => () => Promise.resolve(''));
 
@@ -368,5 +375,71 @@ describe('CLI', () => {
 		const output = process.stdout.write.mock.calls[0][0];
 
 		expect(output).toBe('Custom formatter reports 1 warning(s).');
+	});
+
+	it('--cache same warning config', async () => {
+		await cli([
+			'--cache',
+			'--cache-location',
+			fixturesPath('.stylelintcache'),
+			'--config',
+			fixturesPath('default-severity-warning.json'),
+			fixturesPath('empty-block.css'),
+		]);
+		await cli([
+			'--cache',
+			'--cache-location',
+			fixturesPath('.stylelintcache'),
+			'--config',
+			fixturesPath('default-severity-warning.json'),
+			fixturesPath('empty-block.css'),
+		]);
+		expect(process.stdout.write).toHaveBeenCalledTimes(1);
+
+		await removeFile(fixturesPath('.stylelintcache'));
+	});
+
+	it('--cache different error config', async () => {
+		await cli([
+			'--cache',
+			'--cache-location',
+			fixturesPath('.stylelintcache'),
+			'--config',
+			fixturesPath('default-severity-warning.json'),
+			fixturesPath('empty-block.css'),
+		]);
+		await cli([
+			'--cache',
+			'--cache-location',
+			fixturesPath('.stylelintcache'),
+			'--config',
+			fixturesPath('config-block-no-empty.json'),
+			fixturesPath('empty-block.css'),
+		]);
+		expect(process.stdout.write).toHaveBeenCalledTimes(2);
+
+		await removeFile(fixturesPath('.stylelintcache'));
+	});
+
+	it('--cache same error config', async () => {
+		await cli([
+			'--cache',
+			'--cache-location',
+			fixturesPath('.stylelintcache'),
+			'--config',
+			fixturesPath('config-block-no-empty.json'),
+			fixturesPath('empty-block.css'),
+		]);
+		await cli([
+			'--cache',
+			'--cache-location',
+			fixturesPath('.stylelintcache'),
+			'--config',
+			fixturesPath('config-block-no-empty.json'),
+			fixturesPath('empty-block.css'),
+		]);
+		expect(process.stdout.write).toHaveBeenCalledTimes(2);
+
+		await removeFile(fixturesPath('.stylelintcache'));
 	});
 });
